Remove the status bar overlay (#6613).
Initially, the status bar overlay was used for three purposes: indicate
whether there was a bitmask account configured, as an entry point to the
bitmask account wizard (by clicking on it), and as a watchdog that would popup
the bitmask account wizard everytime thunderbird started, if there was no
account configured.

We don't want any of these features, so this commit completelly removes the
status bar overlay. The issue indicated above (#6613) has to do with removal
of the last of these features (i.e. automated wizard popup).