[feat] Display encryption status when viewing a message
Display a notification bar in the message pane if the message was
encrypted for messages in the "Inbox" folder.

For messages in the "Sent" folder, we have placeholder code for
displaying the message state and for adding custom headers if required.
However, we need to understand how to handle the case for multiple
recipients with different encryption states since it's possible to send
encrypted messages to recipients with known keys but the same message
will be send unencrypted to recipients whose keys are not known. In such
a case we can display the message status per-recipient.