[feat] Display notification to manually fetch messages if Bitmask is shut down
If Bitmask is shut down and Thunderbird is still running, notify the
user using an alert to refresh the session by manually fetching messages
("Get new messages"). Otherwise if Bitmask has shut down and Thunderbird
tries to fetch messages, it will throw up an error and to refresh the
session, the user has to manually fetch the messages.  There can be
better ways of handling this but for now we should just make sure that
the user knows why the messages are not being fetched in case Bitmask
stops running or is shut down.

Another solution can be to check for messages automatically for the user
after Bitmask is running to refresh the session but the reason this
approach was not preferred was because the user may not wish to check
for new messages and force-checking would not be transparent. (Not to
forget, the XUL hacks :)

This commit also removes a redundant string bundle call.